requirements.txt
asgiref==3.6.0
attrs==22.2.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-ckeditor==6.3.2
django-editorjs==0.2.1
django-jalali==6.0.1
django-js-asset==2.0.0
idna==3.4
isodate==0.6.1
jdatetime==4.1.1
lxml==4.9.2
mysqlclient==2.1.1
Pillow==9.5.0
platformdirs==3.2.0
pycryptodome==3.17
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
UNKNOWN==0.0.0
urllib3==1.26.15
zeep==4.2.1
